Make the simulator more flexible
The simulator is made more flexible by storing attributes such as node range in the node instead of in the main simulator class, which allows for a simulation with nodes of different configurations. We also introduce a minimum node range as well as the functionality to spawn in new nodes at runtime. These changes can be seen in ns.py. Since nodes can now have different ranges the connections between them no longer have to be bi-directional, this change is reflected both in ns.py and render.py