Fix bug and improve visuals
Fixed a bug: connections are added to nodes every time create_connections is run but it wouldn't identify duplicates and this caused issues when removing. This is now fixed. Also improved the appearance of connections when rendered, making them easier to identify and tell apart. This is achieved by assigning each connection a random colour and by making the sender's side of the connection thicker than the receivers side, making it possible to see which direction they are facing.